pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "spectre_utils"
version = "0.0.1"
authors = [
  { name="Joe Schoonover", email="joe@fluidnumerics.com" },
]
description = "Tools for setting up ensemble simulations with the MITgcm for the Spectre project"
readme = "README.md"
requires-python = ">=3.11"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: Other",
    "Operating System :: OS Independent",
]

dependencies = [
  "black==25.1.0",
  "bottleneck==1.5.0",
  "cartopy==0.24.1",
  "cdsapi==0.7.6",
  "copernicusmarine==2.1.2",
  "jupyterlab==4.4.3",
  "matplotlib==3.10.3",
  "netCDF4==1.7.4",
  "numpy==2.3.0",
  "pandas==2.3.0",
  "pydap==3.5.8",
  "pyyaml==6.0.1",
  "xarray==2025.6.1",
  "xgcm==0.8.1",
  "zarr==3.0.8",
  "MetPy==1.7.1",
  "claude-agent-sdk",
  "discord.py>=2.3.0",
  "anyio>=4.0.0",
]
[project.urls]
Homepage = "https://github.com/ocean-spectre/spectra-150-ensembles"
Issues = "https://github.com/fluidnumerics/spectre_utils/issues"

[project.scripts]
spectre-agents = "spectre_agents.__main__:cli"

[tool.setuptools]
packages = ["spectre_utils", "spectre_agents", "spectre_agents.tools", "spectre_agents.agents", "spectre_agents.discord_bot"]
